Former-commit-id: 16ad926a78cd0b6a440f720a99557847c219cfbc

+++ Assets/UbiPackages/AssetBundles/Editor/AssetBundleServer/LaunchAssetBundleServer.cs
@@ -75,7 +75,9 @@ namespace AssetBundles
                     return;
 
                 var lastProcess = Process.GetProcessById(instance.m_ServerPID);
-                lastProcess.Kill();
+                if (lastProcess != null)
+                    lastProcess.Kill();
+
                 instance.m_ServerPID = 0;
             }
             catch
